// .metadata/.plugins/org.eclipse.wst.server.core/tmp0/wtpwebapps/com.misys.portal.client.kotak/content/fccui-src/src/app/common/model/corporateDetails.ts
/**
 * Corporate User Profile
 * The Corporate User Profile API provides the information on the corporate banking customers user's: bank,
 * corporate details, corporate entity, user details and the access level privileges.
 *
 * OpenAPI spec version: 1.1.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Address } from './address';
import { References } from './references';


/**
 * Details of the corporate
 */
export class CorporateDetails {
    /**
     * Corporate Id
     */
    id?: string;
    /**
     * Name of the corporate
     */
    name?: string;
    /**
     * Short Name of the corporate
     */
    shortName?: string;
    postalAddress?: Address;
    swiftAddress?: Address;
    references?: References;
    baseCurrency?: any;
    country?: string;
}