// .metadata/.plugins/org.eclipse.wst.server.core/tmp0/wtpwebapps/com.misys.portal.client.kotak/content/fccui-src/src/app/common/model/counterpartyDetailsList.ts
/**
 * Counterparty Maintenance
 * marketing material to be updated
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { CounterpartyDetails } from './counterpartyDetails';
import { Meta } from './meta';


/**
 * View List of counterparties object definition
 */
export interface CounterpartyDetailsList {
    /**
     * List of counterparties
     */
    items?: Array<CounterpartyDetails>;
    meta?: Meta;
}