// .metadata/.plugins/org.eclipse.wst.server.core/tmp0/wtpwebapps/com.misys.portal.client.kotak/content/fccui-src/src/app/common/model/countryList.ts
/**
 * Static Data
 * The Static Data API allows the corporate user to fetch  information on a list of countries and currencies
 *
 * OpenAPI spec version: 1.0.1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Country } from './country';


/**
 * List of Countries Definition
 */
export interface CountryList {
    /**
     * List of Countries
     */
    countries?: Array<Country>;
}